Prevent undefined behavior in getSavedDeviceParams
If there are no save device parameters then calling
cardboard::qrcode::getCurrentSavedDeviceParams() will return a
std::vector of length zero. Immediately following that call the
0th element of the vector is passed to memcpy.

Indexing into a std::vector of zero length results in undefined
behavior. In the implementation of std::vector used by Chrome, for
example, it triggers an assert. This patch checks to see if the
data length is zero and falls back to returning the default
parameters if so to avoid triggering any undefined behaviors.